feat: WIP dowload list of users
