+++ apps/admin/lib/src/features/home/presentation/providers/providers.dart
@@ -1,0 +1,9 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+final getUsersListProvider = FutureProvider(
+  (ref) => FirebaseFirestore.instance
+      .collection('users')
+      .where('email', isNull: false)
+      .get(),
+);
